refactor: 댓글 조회 기능 controller, validator 분리
