+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/social/comment/controller/CommentController.java
@@ -2,21 +2,16 @@ package kakaotech.bootcamp.respec.specranking.domain.social.comment.controller;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
-import kakaotech.bootcamp.respec.specranking.domain.social.comment.dto.CommentListResponse;
 import kakaotech.bootcamp.respec.specranking.domain.social.comment.dto.CommentPostResponse;
 import kakaotech.bootcamp.respec.specranking.domain.social.comment.dto.CommentRequest;
 import kakaotech.bootcamp.respec.specranking.domain.social.comment.dto.CommentUpdateResponse;
 import kakaotech.bootcamp.respec.specranking.domain.social.comment.dto.ReplyPostResponse;
-import kakaotech.bootcamp.respec.specranking.domain.social.comment.service.CommentQueryService;
 import kakaotech.bootcamp.respec.specranking.domain.social.comment.service.CommentService;
 import kakaotech.bootcamp.respec.specranking.global.dto.SimpleResponseDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -32,7 +27,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class CommentController {
 
     private final CommentService commentService;
-    private final CommentQueryService commentQueryService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
@@ -64,12 +58,5 @@ public class CommentController {
             @PathVariable @Positive(message = "스펙 ID는 양수여야 합니다.") Long specId,
             @PathVariable @Positive(message = "댓글 ID는 양수여야 합니다.") Long commentId) {
         return commentService.deleteComment(specId, commentId);
-    }
-
-    @GetMapping
-    public CommentListResponse getComments(
-            @PathVariable @Positive(message = "스펙 ID는 양수여야 합니다.") Long specId,
-            @PageableDefault(size = 5) Pageable pageable) {
-        return commentQueryService.getComments(specId, pageable);
     }
 }
